Changed how input is handled.
Originally, there existed one way to do input. Main would parse a file
and use that information to solve the kenken. Unfortunately, writing a
file that gives all the information asked for by the KenKen solver is a
tedious task. I want to make a move towards a GUI interface, and this was
the first step.

I created an interface called Input that specifies two functions.
public void specifyInput() allows the client to choose from options
acording to the class specific input.
public void getKenKenString() returns the String representation of
the KenKen grid in order to be used for solving.

This interface will hopefully set up the creation of a GUI that
will make the KenKen description less of a hassle to make, by moving
the vessel away from a text file and towards a basic UI.