Separated the solving action from the main.
This is an attempt to set up the interface that will be used with
a GUI I want to write. At the moment, I don't know of a way to have
a main function use JavaFX in an IOC way, so I am having the JavaFX
GUI interact through an interface with the solver.

I needed to first decouple the solving from my main class. I will
probably remove the main class soon, as it can be replaced by the
JavaFX application I plan on writing.