Wrote main as an in-between for InputFromFile and the Solver
Added in an interface called solver that will allow for multiple
ways of solving KenKen to all have the same interface. Currently
I only have the RecursiveSolver, but I want to work on a Solver
that is not recursive.

I changed the Grid interface a little bit. There is no longer a
print function, but instead a getKenKen() function that return
an int[][] that represents the solved (or unsolved) KenKen. Also,
the Grid class now only initializes a Square if it is specified when
constructing a cage.

I also had to remove the validateOptions() functions from the
constructors of the Cages because of a NullPointException that came
from checking valid options before a Grid was completely initialized.