Changed mechanics of making cages
When reading a string to make a cage, instead of using symbols like
+ or / to denote an operation, use the words Add, Divide, etc. This
should make reading the description files a little easier.

I also used this to change the mechanics of the Cage creation. The
use of these words allows me to use the java Reflection API to make
Cages. Instead of using a long if/else block to parse the symbols
and make the corresponding Cage object, I have generic code that
works on all type. This allows the potential to add different types
of boxes later on. Even though kenken is not going to change any
time soon, I thought this would be a nice touch, as it is similar
to how the real world would work, where I don't know what additions
to the code there may be, and I want to write a flexible code as
possible.