Updated JUnit tests, and changed isValidRow function.
JUnit tests now account for the new definition of isValid which
allows for a valid Grid to have Squares whose value is zero.
In order to determine if a Grid is truly valid, one must use
isValid in conjunction with isFilled.

Also changed the isValidRow function to look cleaner. Now uses
a while loop instead of a for loop.