Rewrote isValid and isFilled functions.
isValid and isFilled helper function now have more intuitive
names. Instead of checkRows, as was used in isFilled, there
is isEachRowFilled.

I also rewrote the helper functions for isValid and isFilled to be
more readable. They are now much shorter and concise, and in my
opinion are easier to understand.