Adding a GUI file for the KenKen Solver
Over the past couple of days I have been working on developing the
GUI for the KenKen solver. I am now far enough along that I feel like
it should be added to the Repository. I currently have a Grid visual
with a title. The Squares on the Grid are able to be clicked, and
when clicked have a darkened outline. This will set me up to be able
to group Squares together, and combine them into a Cage.