Finished beginning work on grid visuals, began controls.
The grid visuals will likely be a work in progress throughout this
project because as I add functionality, I'll need to update the
visuals. So I have the basics right now; The grid is always a 5x5
grid and you can click and unclick the Squares.

I added an inner class called VisualCage that will be used to
represent Cages once I move on to that functionality.

I began work on the control panel for the solver gui. I currently
have a drop down selection to choose an operation and a spot to
enter a number for the Cage. Next I will add a button that allows
a user to create a Cage out of the selected Squares.