Finished Cage construction and added an inner Params class
The control panel is now able to combine Squares to make Cages.
The majority of the added logic is in the EventHandler for the
"Create Cage" button, where the drawing also takes place. When the
button is clicked, the handler gets the string value of the drop-
down selection, converts the "Cage Total: " input to an integer,
and builds the Cage. In the process, it redraws the squares that
become the cage in order to mark them separate from the rest of the
grid.

The Params class was added because there were many unqualified numbers
creeping into my code for various properties, mainly widths, heights,
and stroke sizes for various Square states. These are now centralized
in the Params class, which consists of a lot of public static variables.