Currently have Cage binding and solving working.
User can now select sqaures and bind them into Cages, with the
total and operation in the upper-left most Sqaure.

Pressing the solve button takes the information from the VisualCage
class and turns that into String format, which it then passes on to
a Solver (in this case a Recursive solver). Then the button prints
the solved KenKen to System.out. The next step is to have the solve
button also print the solved KenKen onto the KenKen grid.