Solved KenKens output to GUI.
Added in the final functionality to get the GUI working. After
selecting all of the Squares and turning them into Cages, the
solved KenKen is printed onto the grid.

Next steps for this project will be to introduce a method to resize
the grid if the user wants to solve KenKens of a different size. I
also would like to put in comprehensive error reporting.