Removed specifySource() from Input interface.
Removed specifySource() from Input interface because it is primarily
useful only for an input from the command prompt. Considering it
takes no arguments, that would be the primary way one would get
information for the KenKen input. For other applications, the
interface would likely be cumbersome with that method. I also believe
that it is likely better to have the specific class specify the source
for it's KenKen representation inside of its constructor.