Altered the InputFromFile functionality.
Am rethinking if I want the Application to implement the Input
interface. The more I consider how to separate the Application
from the solver, the more difficult it appears. I wanted to try
and multi-thread the two applications, but I run into the issue of
not having a reference to the launched application in the Main thread.
I will give it some thought, but I may just comment my current code,
add in a couple more features, and then move onto another project.