Reorganized gui to be in multiple classes & changed Solver interface to be an abstract class, with the solve method being the abstract method.
Basically a lot of copy and pasting went into this. I moved a lot of
the creation of gui objects out of the KenKenApp.java class and into
their own files. This allows better organization, and I think should
help when I want to add on functionality to the GUI. I did make some
changes to the interface for VisualSquare. VisualSquare also defines
its on action behavior within the class, which is a pattern I adopted
with most of the classes in this commit. This allows each class to
have a well defined default behavior. This behavior can still be
changed with the proper functions, but now the default behavior is
defined with the making of the object, instead of later by the
application.